idea_evaluator_rag/utils/config.py
from dotenv import load_dotenv
import os

load_dotenv()

OLLAMA_URL = os.getenv("OLLAMA_URL", "http://127.0.0.1:11434")
OLLAMA_MODEL = os.getenv("OLLAMA_MODEL", "mistral:latest")
CHROMA_DB_DIR = os.getenv("CHROMA_DB_DIR", "./chroma_store")
